+++ vhost-device-console/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### Added
 
+- [#811](https://github.com/rust-vmm/vhost-device/pull/811) Being able to specify max queue size
+
 ### Changed
 
 ### Fixed